Revert "Starting Refactor of Account.java"
